Factor out storefinder from Pet Valu
